Major revision
To address the discussion in #1, the ATS and the conformance classes
have been revised completely. To illustrate and test the approach, ETSs
have been developed for some of the conformance classes. A small sample
test data set has been accessed from an INSPIRE download service and
tested against the ETSs. The resulting test report is also added.